+++ packages/@githon/asciidoc/README.md
@@ -1,0 +1,3 @@
+# GitBook AsciiDoc Parser
+
+This node module uses AsciiDoctor to parse AsciiDoc for gitbook (SUMMARY.adoc, README.adoc).
