+++ packages/@githon/asciidoc/lib/index.js
@@ -1,0 +1,5 @@
+const HTMLParser = require("gitbook-html");
+const toHTML = require("./toHTML");
+const toAsciidoc = require("./toAsciidoc");
+
+module.exports = HTMLParser.createParser(toHTML, toAsciidoc);
