+++ packages/@githon/asciidoc/test/fixtures/GLOSSARY.adoc
