+++ packages/@githon/html/README.md
@@ -1,8 +1,5 @@
 # GitBook HTML Parser
 
-[![Build Status](https://travis-ci.org/GitbookIO/gitbook-html.png?branch=master)](https://travis-ci.org/GitbookIO/gitbook-html)
-[![NPM version](https://badge.fury.io/js/gitbook-html.svg)](http://badge.fury.io/js/gitbook-html)
-
 This node module parse HTML for gitbook (SUMMARY.html, README.html).
 
 It is a dependency of **gitbook-asciidoc** and **gitbook-markdown**.
