+++ packages/@githon/markdown/README.md
@@ -1,6 +1,3 @@
 # GitBook Markdown Parser
 
-[![Build Status](https://travis-ci.org/GitbookIO/gitbook-markdown.png?branch=master)](https://travis-ci.org/GitbookIO/gitbook-markdown)
-[![NPM version](https://badge.fury.io/js/gitbook-markdown.svg)](http://badge.fury.io/js/gitbook-markdown)
-
 This node module uses kramed to parse markdown for gitbook (SUMMARY.md, README.md).
