--- packages/gitbook-asciidoc/.travis.yml
@@ -1,4 +1,0 @@
-language: node_js
-node_js:
-  - "stable"
-  - "0.12"
