--- packages/gitbook-asciidoc/lib/index.js
@@ -1,5 +1,0 @@
-var HTMLParser = require('gitbook-html');
-var toHTML = require('./toHTML');
-var toAsciidoc = require('./toAsciidoc');
-
-module.exports = HTMLParser.createParser(toHTML, toAsciidoc);
