--- packages/gitbook-asciidoc/package.json
@@ -1,37 +1,0 @@
-{
-    "name": "gitbook-asciidoc",
-    "version": "1.2.2",
-    "homepage": "https://www.gitbook.com",
-    "description": "Parse AsciiDoc content for gitbook",
-    "main": "lib/index.js",
-    "dependencies": {
-        "lodash": "^4.13.1",
-        "asciidoctor.js": "1.5.5-4",
-        "gitbook-html": "1.3.3"
-    },
-    "devDependencies": {
-        "mocha": "2.3.2"
-    },
-    "scripts": {
-        "test": "export TESTING=true; mocha --reporter list --bail"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/GitbookIO/gitbook-asciidoc.git"
-    },
-    "author": "FriendCode Inc. <contact@gitbook.com>",
-    "license": "Apache-2.0",
-    "bugs": {
-        "url": "https://github.com/GitbookIO/gitbook-asciidoc/issues"
-    },
-    "contributors": [
-        {
-            "name": "Aaron O'Mullan",
-            "email": "aaron@gitbook.com"
-        },
-        {
-            "name": "Samy Pessé",
-            "email": "samy@gitbook.com"
-        }
-    ]
-}
